background info: the wire protocol here uses commas to seperate the lines of output it gets from machtype while passing the info from the client to the server.
1) sun processors are reported as SUNW,foo.  Fixed by looking for
those 5 characters, and skipping over them.

2) SGIs report 3 lines of processor info.  Look for and skip the 2nd
and third ones.

3) The code to translate the display code was only looking at the
first word on any line.  I nuked the code that was going out of my way
to do this.

4) The SGI display info includes multiple lines, and a gazillion
commas too boot.  I just nuked the stuff that looped on all of the
lines of display.  No other platform produces multiple lines, not even
if they happen to have multiple displays.

5) The clients have been passing the Athena version to the server
since at least 7.7, and possibly as far back as 7.3, yet the server
had no idea it should expect it.